requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.23.0
Pillow>=9.0.0

